Extract HelpContact; add to LaunchError page
