+++ ccm_web/client/src/pages/LaunchError.tsx
@@ -2,10 +2,13 @@ import React from 'react'
 
 import { Typography } from '@material-ui/core'
 
+import HelpContact, { HelpContactProps } from '../components/HelpContact'
 import InlineErrorAlert from '../components/InlineErrorAlert'
 import Layout from '../components/Layout'
 
-export default function LaunchError (): JSX.Element {
+interface LaunchErrorProps extends HelpContactProps {}
+
+export default function LaunchError (props: LaunchErrorProps): JSX.Element {
   return (
     <Layout>
       <InlineErrorAlert>
@@ -14,6 +17,7 @@ export default function LaunchError (): JSX.Element {
           The tool could not be launched. It is likely your browser is blocking third-party cookies,
           which are required for this application.
         </Typography>
+        <HelpContact {...props} />
       </InlineErrorAlert>
     </Layout>
   )
